+++ src/main/java/com/example/moizaspringserver/domain/feed/respository/FeedAttachmentFileRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.feed.respository;
+
+import com.example.moizaspringserver.domain.feed.entity.FeedAttachmentFile;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FeedAttachmentFileRepository extends CrudRepository<FeedAttachmentFile, Integer> {
+}
