+++ src/main/java/com/example/moizaspringserver/domain/feed/respository/FeedRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.feed.respository;
+
+import com.example.moizaspringserver.domain.feed.entity.Feed;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FeedRepository extends JpaRepository<Feed, Integer> {
+}
