+++ src/main/java/com/example/moizaspringserver/domain/feed/respository/LocalFeedRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.feed.respository;
+
+import com.example.moizaspringserver.domain.feed.entity.LocalFeed;
+import org.springframework.data.repository.CrudRepository;
+
+public interface LocalFeedRepository extends CrudRepository<LocalFeed, Integer> {
+}
