+++ src/main/java/com/example/moizaspringserver/domain/feed/respository/PublicFeedRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.feed.respository;
+
+import com.example.moizaspringserver.domain.feed.entity.PublicFeed;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PublicFeedRepository extends JpaRepository<PublicFeed, Integer> {
+}
