+++ src/main/java/com/example/moizaspringserver/domain/feed/type/FeedType.java
@@ -1,0 +1,6 @@
+package com.example.moizaspringserver.domain.feed.type;
+
+public enum FeedType {
+    QUESTION,
+    COMMON
+}
