+++ src/main/java/com/example/moizaspringserver/domain/follow/repository/FollowRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.follow.repository;
+
+import com.example.moizaspringserver.domain.follow.entity.Follow;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FollowRepository extends CrudRepository<Follow, Integer> {
+}
