+++ src/main/java/com/example/moizaspringserver/domain/notification/repository/NotificationEntityRepository.java
@@ -1,0 +1,7 @@
+package com.example.moizaspringserver.domain.notification.repository;
+
+import com.example.moizaspringserver.domain.notification.entity.NotificationEntity;
+import org.springframework.data.repository.CrudRepository;
+
+public interface NotificationEntityRepository extends CrudRepository<NotificationEntity, Integer> {
+}
