+++ src/main/java/com/example/moizaspringserver/domain/user/repository/UserRepository.java
@@ -3,5 +3,5 @@ package com.example.moizaspringserver.domain.user.repository;
 import com.example.moizaspringserver.domain.user.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 }
