--- src/main/java/com/example/moizaspringserver/feed/respository/FeedAttachmentFileRepository.java
@@ -1,7 +1,0 @@
-package com.example.moizaspringserver.feed.respository;
-
-import com.example.moizaspringserver.feed.entity.FeedAttachmentFile;
-import org.springframework.data.repository.CrudRepository;
-
-public interface FeedAttachmentFileRepository extends CrudRepository<FeedAttachmentFile, Long> {
-}
