--- src/main/java/com/example/moizaspringserver/feed/respository/LocalFeedRepository.java
@@ -1,7 +1,0 @@
-package com.example.moizaspringserver.feed.respository;
-
-import com.example.moizaspringserver.feed.entity.LocalFeed;
-import org.springframework.data.repository.CrudRepository;
-
-public interface LocalFeedRepository extends CrudRepository<LocalFeed, Long> {
-}
