--- src/main/java/com/example/moizaspringserver/feed/respository/PublicFeedRepository.java
@@ -1,7 +1,0 @@
-package com.example.moizaspringserver.feed.respository;
-
-import com.example.moizaspringserver.feed.entity.PublicFeed;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PublicFeedRepository extends JpaRepository<PublicFeed, Long> {
-}
