--- src/main/java/com/example/moizaspringserver/feed/type/FeedType.java
@@ -1,6 +1,0 @@
-package com.example.moizaspringserver.feed.type;
-
-public enum FeedType {
-    QUESTION,
-    COMMON
-}
