+++ src/main/java/com/example/moizaspringserver/global/entity/BaseIdEntity.java
@@ -13,6 +13,6 @@ public abstract class BaseIdEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
 }
