+++ src/main/java/com/example/moizaspringserver/global/entity/BaseTimeIdEntity.java
@@ -13,6 +13,6 @@ public abstract class BaseTimeIdEntity extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
 }
